ni-zynq: list NI UARTs as compatible with ns16550a
Currently, when booting upstream kernels that do not have support for NI
UARTs, you can specify the NI UARTs in the device tree as "ns16550a" in
order to get a working serial console.

To make this easier, list "ns16550a" as a secondary option under the
'compatible' tag for the NI UARTs in the device tree.  That way, the
kernel will use the NI-specific driver if it is available, but will fall
back to the generic driver otherwise.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Joe Hershberger <joseph.hershberger@ni.com>
Reviewed-by: Jaeden Amero <jaeden.amero@ni.com>
---
Natinst-ReviewBoard-ID: 96055