nati_zynq_defconfig: Remove user-mode access to XADC
If any bitfile is loaded to a Zynq target, accessing the user-mode XADC
handles results in an imprecise external abort in the kernel. Since the
only product that actively uses the XADC (roboRIO) uses it through the
FPGA and disables the user-mode module anyway, disable for all other
devices until we can determine how to fix in a future release.

Signed-off-by: Xander Huff <xander.huff@ni.com>
Reviewed-by: Jaeden Amero <jaeden.amero@ni.com>
--
Natinst-CAR-ID: 527175
Natinst-ReviewBoard-ID: 97916