nati_zynq_defconfig: disable PL310_ERRATA_588369
The PL310 errata 588369 is fixed in revision r2p0.

The Zynq chips we are using are documented as having revision
r3p2-50rel0 of the PL310, so this errata is not relevant on our
platforms.  Disable it.

Suggested-by: Gratian Crisan <gratian.crisan@ni.com>
Signed-off-by: Josh Cartwright <joshc@ni.com>
(cherry picked from commit 9e2a5f4f128bc13c34f273cac6331e8aa81c64b2)
Signed-off-by: deooi <deborah.ooi@ni.com>