e1000e: Add delays after writing to registers
There is a noticeable impact on determinism when a large number of
writes are flushed. Writes to the hardware registers are sent across
the PCI bus and take a significant amount of time to complete after
a flush, which causes high priority tasks (including interrupts) to
be delayed. This problem stems from an issue with the PCI bus where
a fabric lock is held during I/O buffer drain. This process can be
detrimental to real-time systems and is seen whenever a large number
of MMIO writes are issued. In the case of the e1000 drivers, when a
device is reset several tables (MTA, VLAN, etc) are rewritten,
generating enough MMIO writes over PCI for the latency issues to be
seen.

Adding a delay after long series of writes gives them time to
complete (drain the I/O buffer), and for higher priority tasks to
run unimpeded.

Signed-off-by: Jonathan David <jonathan.david@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
[bstreiff: remove unused 'rar_count']
Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
[gratian: update Kconfig '---help---' to 'help']
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>