Xilinx UDC: Don't get confused by VBUS Valid
[backport of 3.14 commit e7c3c0a4c9cc79218cdb655368ac1b7a9f05efac]

The Zynq USB MAC, when used in device mode with an ULPI PHY, will get
confused and stop working if it receives some VBUS Valid interrupts,
even when the MAC is not configured to pay attention to VBUS Valid
interrupts. To work around this problem, we disable VBUS Valid
interrupts on the PHY through the ULPI Viewport at setup time. The MAC
never enables or disables the VBUS Valid interrupt when used in device
mode, so no fighting over the PHY interrupt configuration will occur.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Jeff Westfahl <jeff.westfahl@ni.com>
Acked-by: Joe Hershberger <joe.hershberger@ni.com>
[brandon.streiff@ni.com: zynq_writel is xusbps_writel in this version]
Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>