rtmutex: Fix check in __try_to_take_rt_mutex()
In reboot testing, a kernel OOPs was seen as a result of
attempting to reference a NULL waiters_leftmost of a lock
rt_spin_lock_slowlock->
  __try_to_take_rt_mutex->
    rt_mutex_top_waiter

Consider the following scenario:
task A blocks on rt_mutex L1, owned by task B.
A is scheduled in, attempts to take L1
rt_spin_lock_slowlock(L1)->
    raw_spin_lock(L1->wait_lock)

    // Create waiter W1 incase we need to queue on lock
    __try_to_take_rt_mutex(L1, A, NULL,...)

    // enqueue-less attempt fails, enqueue/boost
    task_blocks_on_rt_mutex(L1, W1, A,...)

    for(;;) {
        __try_to_take_rt_mutex(L1, A, W1,...)
        // Second attempt fails
        raw_spin_unlock(L1->wait_lock)
        schedule_rt_mutex(L1) // schedule()

At this point, something (task B?) removes W1 from L1's waiters rbt.

Add in a simple sanity check that there are indeed waiters on a lock
prior to attempting to get the top waiter. The subsequent dequeuing
of a waiter from an empty rbt is a noop.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>