Revert "usb: remove unnecessary USB_OTG dependency on power management"
This reverts commit 2d3e63dce2ee1939aae7f78604fbe7705fad7810.

This was done to support the USB_OTG on Smasher which does not have power
management enabled. But we don't need this because USB gadget support is
not dependent on USB_OTG and we do not need it in our defconfig.
We will be removing the USB_OTG support from our defconfig.

Signed-off-by: Rishika Gupta <rishika.gupta@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Josh Cartwright <josh.cartwright@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Xander Huff <xander.huff@ni.com>
Natinst-ReviewBoard-ID: 118133