irqchip: gic: Add legacy irqdomain option for GIC
We have drivers (like NI-RIO) that contain hardcoded IRQ numbers, and
thus rely on Linux having a stable map of virtual IRQ number to hardware
IRQ number.  Using a linear IRQ domain will break that, so keep using a
legacy irqdomain instead of a linear one when GIC_LEGACY_IRQDOMAIN is set.

This is a non-upstreamable HACK to maintain compatibility with NI-RIO.

Signed-off-by: Nathan Sullivan <nathan.sullivan@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Xander Huff <xander.huff@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>

Natinst-CAR-ID: 567714
Natinst-ReviewBoard-ID: 123493