ARM: dts: ni-793x: split out separate device trees for each variant
Splitting out each of the flexrio variants is a preparation step for
renaming the NI dtbs to include a device code (to simplify build
process, and aide packaging).

Using dtc to decompile both before and after this change, I was able to
confirm that the only property which changed in the resulting dtbs was
the model code.

Signed-off-by: Josh Cartwright <joshc@ni.com>