HACK: ath6kl: Reset the radio when it is powered down
On cRIO controllers, we use the Intel SDIO controller which does
not have the SD/SDIO power control pin connected as required by
the SD card spec.  This is an errata on their part.  There is no
plumbing in the SD Host controller code for this for any PCI
devices.

Work around this by twiddling the radio's reset line when it should
be powered down.

Signed-off-by: James Minor <james.minor@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>

Natinst-ReviewBoard-ID: 130170
Natinst-CAR-ID: 566029