ni-*dts: Add 802.3 compatibility string
As of 801a8ef54 (of: phy: Only register a phy device for phys),
the OF MDIO code checks for either some 802.3 compatible strings or
no compatible strings at all. This resulted in the kernel thinking
that the PHYs in the NI DTs (all of which call out a compatibility
property) were something other than PHYs.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>