HACK: phy: marvell: Add no-op 88E151x soft_reset
930b37ee (net: phy: Add SGMII support for Marvell 88E1510/1512/1514/1518)
introduced a config_init function for the driver. This leads to the
following behavior for DT-defined marvell phys:

  of_phy_connect()
    -> phy_connect_direct()
      -> phy_attach_direct()
        -> phy_init_hw()

In phy_init_hw, if there's a driver defined for the phy that
defines a config_init function, there is a call to the driver's
soft_reset() (or the genphy version if the driver doesn't provide
one).

Prior to this call, interrupts may have been configured. Calling
the genphy_soft_reset causes the phy interrupt config to be applied,
resulting in interrupts prior to the netdev being up.

Provide a no-op reset to maintain existing behavior (the later call
to the driver-specific config_init() in phy_init_hw() ensures that
SGMII setup or DT-provided register setup occurs, otherwise is a
no-op). This prevents interrupts from firing prior to the netdev
being prepared.

Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Josh Cartwright <joshc@ni.com>

Natinst-ReviewBoard-ID: 148256