HACK: net: macb: Trigger phy state machine on _open()
d5c3d8465 (net: phy: Avoid polling PHY with PHY_IGNORE_INTERRUPTS)
changed the behavior of the phy state machine to only be triggered
on interrupts.

Previously, the state machine work would re-arm itself regardless
of whether the phydev supported interrupts.

With the state machine being run to update state only on phy irqs
when dealing with an irqful phy, if that phy were to interrupt early
(before the macb ndev is open()ed) or only after entering aneg,
the phy will remain in UP as there's nothing to run phy_state_machine
(it either already fired interrupts or the phy is waiting to enter
ANEG to make changes that will fire interrupts)

For a more general solution, force the state machine to rerun
after starting the mac (moving the phy state machine from READY
to UP, with nothing to move into AN).

Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Josh Cartwright <joshc@ni.com>

Natinst-ReviewBoard-ID: 148264