nati_*_defconfig: disable BTRFS
AFAIK no one is using BTRFS on nilrt; it was enabled along with many
others in 79ec8745992 ("nati_zynq_defconfig: Regenerate, add
functionality").

Between the 2016 and 2017 releases the kernel increased around 5.8
mb which is problematic for the memory constrained devices running
in safemode with all kernel modules installed by default. BTRFS is
about 1.3 mb (including it's raid6_pq dependency) out of those 5.6
so remove it to free up disk space.

Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Ken Sharp <ken.sharp@ni.com>
Acked-by: Bill Pittman <bill.pittman@ni.com>
Acked-by: Rich Tollerton <rich.tollerton@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Natinst-ReviewBoard-ID: 168207