i2c: xiic: Handle devices that do not specify a clock
With the addition of commit e9aacb08ff9d ("i2c: xiic: Implement power
management") the xiic driver failed to load if a clock was not specified
in the dts.

This commit changes the behavior so that a clock does not need to be
specified in order for the driver to function.

Fixes: e9aacb08ff9d ("i2c: xiic: Implement power management")
Signed-off-by: Zach Brown <zach.brown@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Kyle Roeschley <kyle.roeschley@ni.com>
Acked-by: Julia Cartwright <julia@ni.com>
Acked-by: Brandon Streiff <brandon.streiff@ni.com>
Natinst-CAR-ID: 630351
Natinst-ReviewBoard-ID: 173828