usb: dwc3: gadget: request interrupt with IRQF_NO_THREAD
Commit 2a1d3ab8986d ("genirq: Handle force threading of irqs with primary
and thread handler") enables force threading of irq primary and secondary
handlers.

The maintainer for the dwc3/gadget subsystem, Felipe Balbi, discussed its
effect on this driver with Thomas Gleixner the author of the patch listed
above here: https://lkml.org/lkml/2015/10/9/379

The conclusion reached was to use the IRQF_NO_THREAD and keep the primary
handler in interrupt context, however a patch was never applied.

In addition to the performance concerns described in the mailing list post,
we have discovered that force threading the primary handler for dwc3/gadget
driver introduces a race condition between the primary and the secondary
threaded handlers when used with PREEMPT_RT_FULL.

The primary handler in the dwc3 gadget driver is used only to detect what
device generated the IRQ and based on that wake up the threaded handler.
Also IRQF_ONESHOT is not used because the IRQs for this device are masked
in the primary handler and only unmasked in the threaded secondary handler.
Based on this, the use of IRQF_NO_THREAD is the appropriate solution here
so implement it.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Xander Huff <xander.huff@ni.com>
Natinst-CAR-ID: 623246
Natinst-ReviewBoard-ID: 174515