net: ath6kl: Ignore error from sdio_disable_func
The ath6kl has hardware problems. When the driver detects something has
gone wrong it attempts to turn the device off and on again.
As part of that process, sdio_disable_func is called.
commit 5d3f6ef04d6 ("mmc: Wait for card_busy before starting sdio
requests") changes the behavior of functions that call mmc_io_rw_direct,
which sdio_disable_func does. Now if the card is 'busy' mmc_io_rw_direct
will attempt to wait until isn't or will eventually timeout. This causes
sdio_disable_func to fail when the ath6kl is misbehaving, which in turn
aborts the process of resetting ath6kl.

This commit fixes the issue by simply ignoring the result of
sdio_disable_func and moving on.

Signed-off-by: Zach Brown <zach.brown@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Fixes: 5d3f6ef04d6 ("mmc: Wait for card_busy before starting sdio 10 requests")
Acked-by: Kyle Roeschley <kyle.roeschley@ni.com>
Acked-by: James Minor <james.minor@ni.com>
Natinst-CAR-ID: 681138
Natinst-ReviewBoard-ID: 219783