Revert "mmc: core: enable CMD19 tuning for DDR50 mode"
This reverts commit 4324f6de6d2eb9b232410eb0d67bfafdde8ba711.

The SD Physical Layer Simplified Specification Version 6.00 §4.2.4.5 says
that tuning is only supported for SDR50 and SDR104 modes, not DDR50 mode.
As a result, tuning always fails for DDR50 cards. Remove "support" for this
in the kernel, as it only adds unnecessary error prints.

This was tested with a variety of industrial and consumer microSD cards
from Swissbit, Unirex, Patriot, SanDisk, and Link Depot. None of the cards
were able to execute tuning in DDR50 mode.

Signed-off-by: Kyle Roeschley <kyle.roeschley@ni.com>
Signed-off-by: Brad Mouring <brad.mouring@ni.com>
Acked-by: Tony Liechty <tony.liechty@ni.com>
Acked-by: Nathan Sullivan <nathan.sullivan@ni.com>
Natinst-ReviewBoard-ID: 236135