HACK: mtd: nand: Don't bail out early if on-die ECC can't be disabled
Commit 9748e1d87573 ("mtd: nand: add support for Micron on-die ECC")
added support for Micron on-die ECC, but is missing support for chips
that forcefully enable on-die ECC. As a result it deliberately bails
out if the ECC cannot be disabled.

This is a temporary commit to get things booting. Replace with proper
support for Micron NANDs that do not allow disabling on-die ECC.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>