usb: dwc3: pci: Fix property init for Apollo Lake
The initialization of device properties for Apollo Lake and Broxton CPUs was
bailing out early which prevented the definition of sysdev_is_parent and
dr_mode properties.

The missing sysdev_is_parent property made a call to dma_alloc_coherent() fail
when a USB gadget driver was loaded.

The missing dr_mode property lead to a non-fatal dmesg error.

To fix this, ensure that the sysdev_is_parent and dr_mode property is defined
for Apollo Lake and Broxton CPUs.

Fixes: 2012c036bf28 ("usb: dwc3: call _DSM for core soft reset")
Signed-off-by: Ken Sharp <ken.sharp@ni.com>
Acked-by: Kyle Roeschley <kyle.roeschley@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Natinst-CAR-ID: 715260
Natinst-ReviewBoard-ID: 259661