mmc: sdhci: Do not disable interrupts in sdhci_set_clock
Upstream removed spin lock usage in the set_ios path with commit
d1e4f74f911d ("mmc: sdhci: Do not use spin lock in set_ios paths"), which
means that our calls to spin_(un)lock_irq() in sdhci_set_clock() now cause
an error on boot if a card is not present or a crash if one is present.
Remove these calls so we don't do that and to match upstream's change.

Fixes: 815777e0a1fb2 ("mmc: sdhci: Add quirk for delay between clock disable and param change")
Fixes: 30d2f458ba48f ("mmc: sdhci: Add quirk work around double rescan")

Signed-off-by: Kyle Roeschley <kyle.roeschley@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Brandon Streiff <brandon.streiff@ni.com>
Natinst-ReviewBoard-ID: 264712
Natinst-CAR-ID: 720310