i2c: cadence: Fix the hold bit setting on transfer
In the case where a transfer doesn't use the HOLD setting, we incorrectly
carry over the last HOLD value that was used. This means that any transfer
that doesn't use HOLD which immediately follows a transfer that used HOLD
will fail with a bus timeout. Fix this by clearing the bit when it isn't
needed.

This fixes sporadic i2c bus lockups on National Instruments Zynq-based
devices.

Fixes: df8eb5691c48 ("i2c: Add driver for Cadence I2C controller")
Acked-by: Michal Simek <michal.simek@xilinx.com>
Reported-and-tested-by: Kyle Roeschley <kyle.roeschley@ni.com>
Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@xilinx.com>
Natinst-CAR-ID: 720112
Natinst-ReviewBoard-ID: 275215