genirq: Fix canceling the old_notify work in irq_set_affinity_notifier()
Commit 3e4242082f03 ("genirq: Handle missing work_struct in
irq_set_affinity_notifier()") introduced a special case to handle the
fact that cancel_work_sync() on a work_struct element is not available
in RT. However it erroneously tries to cancel the new 'notify' work
instead of the 'old_notify' work in the second part of
irq_set_affinity_notifier(). This results in a NULL pointer
dereference when irq_set_affinity_notifier() is called with a NULL
'notify' value which is the case when trying to disable a
notification.

Fix this by making the clean-up path of irq_set_affinity_notifier()
use 'old_notify' consistently.

Fixes: 3e4242082f03 ("genirq: Handle missing work_struct in irq_set_affinity_notifier()")
Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Reported-by: Kar Hin Ong <kar.hin.ong@ni.com>
Natinst-AZDO-ID: 952207
Natinst-ReviewBoard-ID: 321414
Acked-by: Kar Hin Ong <kar.hin.ong@ni.com>