g_lci: make mass storage and rndis interfaces runtime-selectable
Allow the mass storage and RNDIS interfaces to be selected via module
parameters. This lets us select them from the kernel cmdline, which
means that we can service all firmware builds with a single kernel
configuration.

For mass storage, we key off of the existing 'g_lci.cdrom=1' param,
the reasoning being that we wouldn't expose a read-write volume anyway.
For RNDIS, a new 'g_lci.rndis' param is added.

In a much-more-post-3.2 world, this could all be configured from
userspace via libcomposite, but sometimes you have to go into reviving
old products with the kernel you have, not the kernel you want...

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>