nati_zynq_defconfig: Fix typo error in commit 9e8c398d23b6
Commit 9e8c398d23b6 contained 2 typos. The "_FS_" portions for
"CONFIG_EXT4_FS_POSIX_ACL" and "CONFIG_EXT4_FS_SECURITY" were missing.
This caused the config to not reflect properly in the kernel which
broke the software installation and other services like
systemWebServer that depends on the configs.

Also, "CONFIG_EXT3_FS=y" is not needed anymore because it is replaced
by EXT4 drivers and is backward compatible.

This commit corrects the typos and removes "CONFIG_EXT3_FS=y".

Signed-off-by: wkoe <wilson.koe@ni.com>